Crawling and Data Processing
We need to visualize the impact by global Warming.
The Sea Surface Temperatue Change is very effective to visualize this.

So, we obatain The Temperature data near korea using 'python jsoup crawling'.
we process the data and plot 2D-Temperature by X,Y Loaction .
last we compare trend the fish type and temperature variation.